moved the game into own path to not make everypath client
